refactor: move getFileSizeToClosestByte to root utils
